max speed and accel control changes
- Unified speed across modes
- Changed calibration from long press to double-tap
- Allowed switching to control accel by tapping encoder

Unfortunately this scraps some work that Jacob just did to fix
my broken implementation of speed percentages. There's parts of
that that would be nice to get into here if we have time, like
adding a bit of slop to either side of the maximum, so that if a
user turns past max, it doesn't _immediately_ pop back to one
tick below max when they turn it one tick back (hopefully that
makes sense.) Unfortunately it seemed a bit too difficult to just
adapt the speed percentage code to our purpose, since we need to
be able to gracefully adjust to new accel values coming in from
Serial from the app.